+++ src/main/java/uz/mu/autotest/extractor/impl/JavaTestResultParser.java
@@ -5,16 +5,16 @@ import jakarta.xml.bind.JAXBException;
 import jakarta.xml.bind.Unmarshaller;
 import org.springframework.stereotype.Component;
 import uz.mu.autotest.extractor.Parser;
-import uz.mu.autotest.extractor.java.TestSuite;
+import uz.mu.autotest.extractor.java.JavaTestSuite;
 
 import java.io.File;
 
 @Component
-public class JavaTestResultParser implements Parser<TestSuite> {
+public class JavaTestResultParser implements Parser<JavaTestSuite> {
     @Override
-    public TestSuite parse(File file) throws JAXBException {
-        JAXBContext context = JAXBContext.newInstance(TestSuite.class);
+    public JavaTestSuite parse(File file) throws JAXBException {
+        JAXBContext context = JAXBContext.newInstance(JavaTestSuite.class);
         Unmarshaller unmarshaller = context.createUnmarshaller();
-        return (TestSuite) unmarshaller.unmarshal(file);
+        return (JavaTestSuite) unmarshaller.unmarshal(file);
     }
 }
