+++ src/main/java/uz/mu/autotest/extractor/impl/PythonTestResultParser.java
@@ -5,15 +5,15 @@ import jakarta.xml.bind.JAXBException;
 import jakarta.xml.bind.Unmarshaller;
 import org.springframework.stereotype.Component;
 import uz.mu.autotest.extractor.Parser;
-import uz.mu.autotest.extractor.python.TestSuite;
+import uz.mu.autotest.extractor.python.PythonTestSuite;
 import uz.mu.autotest.extractor.python.TestSuites;
 
 import java.io.File;
 
 @Component
-public class PythonTestResultParser implements Parser<TestSuite> {
+public class PythonTestResultParser implements Parser<PythonTestSuite> {
     @Override
-    public TestSuite parse(File file) throws JAXBException {
+    public PythonTestSuite parse(File file) throws JAXBException {
         JAXBContext context = JAXBContext.newInstance(TestSuites.class);
         Unmarshaller unmarshaller = context.createUnmarshaller();
         TestSuites testSuites = (TestSuites) unmarshaller.unmarshal(file);
